Changes to the trade function
Switched to a more appropriate way to take profits by using  a conditional algo-order instead of a market buy and a limit sell order. Although it doesn't seem to work because the takeprofit trigger price is somehow invalid. Too confused to fix this rn.

(im losing my sanity pls help)